+++ src/main/java/com/hyewon/springplayground/kafka/KafkaListenerService.java
@@ -20,7 +20,6 @@ public class KafkaListenerService {
     @Autowired
     private ConcurrentKafkaListenerContainerFactory<String, String> kafkaListenerContainerFactory;
 
-
     public void stopListener() {
         MessageListenerContainer listenerContainer = registry.getListenerContainer("myListener");
         if (listenerContainer != null) {
