+++ src/main/resources/application.yml
@@ -54,5 +54,10 @@ server:
 
 kakao:
   client:
-    id: 20c5e61eca234407fe67e0918dbff845
-    secret: LNb25bRip0Xxa8utallgJUTz1ycaWs06
+    id: 1
+    secret: 1
+
+#otel:
+#  exporter:
+#    otlp:
+#      endpoint: http://localhost:4318
